requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
numpy>=1.24.0
faiss-cpu>=1.7.4
requests>=2.31.0
beautifulsoup4>=4.12.0
tiktoken>=0.7.0
yfinance>=0.2.30
crewai>=0.30.0
ollama>=0.3.0
langchain-openai>=0.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
typing_extensions>=4.10.0
